+++ src/components/Avatar/index.ts
@@ -1,3 +1,3 @@
-// Avatar from copy-ui @ 2025-11-18.
+// Avatar from copy-ui @ 2025-11-19.
 
 export { Avatar } from "./avatar";