+++ src/components/Avatar/CHANGELOG.md
@@ -6,3 +6,5 @@
   And add a border.
 - 2025-11-18: Add `Avatar.Fallback` sub-component and support to show it when
   `Avatar.Img` not loaded.
+- 2025-11-19: Use `useRef` to hold the avatar store instance. Add make `Avatar`
+  expose more props as a `span` element.